+++ app/controllers/api/v1/fleet_members_controller.rb
@@ -43,7 +43,7 @@ module Api
 
       private def member_query_params
         @member_query_params ||= query_params(
-          :username_cont, sorts: [], role_in: []
+          :username_cont, :name_cont, sorts: [], role_in: []
         )
       end
     end