+++ app/controllers/api/v1/stations_controller.rb
@@ -64,7 +64,7 @@ module Api
       private def station_query_params
         @station_query_params ||= query_params(
           :with_shops, :celestial_object_eq, :name_cont, :slug_eq, :habs_not_null, :search_cont,
-          :commodity_item_type,
+          :commodity_item_type, :cargo_hub_eq, :refinery_eq,
           name_in: [], celestial_object_in: [], starsystem_in: [], station_type_in: [],
           shops_shop_type_in: [], docks_ship_size_in: [], sorts: []
         )